+++ resources/js/components/Room/MembersComponent.vue
@@ -39,20 +39,29 @@
           :per-page="settings('pagination_page_size')"
           :fields="tableFields"
           :items="members"
+          v-model="displayedMembers"
           hover
+          ref="membersTable"
           stacked="lg"
           show-empty
+          selectable
+          :no-select-on-click="true"
+          @row-selected="onRowsSelected"
         >
           <!-- message on no members -->
           <template v-slot:empty>
             <i>{{ $t('rooms.members.nodata') }}</i>
           </template>
 
-          <!-- table data fetching spinner -->
-          <template v-slot:table-busy>
-            <div class="text-center my-2">
-              <b-spinner class="align-middle"></b-spinner>
-            </div>
+          <!-- checkbox in header -->
+          <template #head(selected)>
+            <b-form-checkbox v-if="selectableMembers>0" :checked="selectedMembers.length == selectableMembers && displayedMembers.length > 0" @change="onAllRowsSelected" />
+            <span v-else></span>
+          </template>
+
+          <!-- checkbox to select the current row -->
+          <template #cell(selected)="{ item, rowSelected, selectRow, unselectRow }">
+            <b-form-checkbox v-if="currentUser && currentUser.id !== item.id" :checked="rowSelected" @change="rowSelected ? unselectRow() : selectRow() "/>
           </template>
 
           <!-- action buttons -->
@@ -76,7 +85,7 @@
                 @click="showRemoveMemberModal(data.item)"
                 v-b-tooltip.hover
                 v-tooltip-hide-click
-                :title="$t('rooms.members.deleteUser')"
+                :title="$t('rooms.members.removeUser')"
               >
                 <i class="fa-solid fa-trash"></i>
               </b-button>
@@ -106,13 +115,37 @@
         <b-row>
           <b-col cols="12" class="my-1">
             <b-pagination
-              v-if="members.length>settings('pagination_page_size')"
               v-model="currentPage"
               :total-rows="members.length"
               :per-page="settings('pagination_page_size')"
             ></b-pagination>
           </b-col>
         </b-row>
+        <!-- selected rows action buttons -->
+        <b-button-group class="float-md-right" v-if="selectedMembers.length > 0">
+            <!-- bulk edit membership role -->
+            <b-button
+              :disabled="isBusy"
+              variant="dark"
+              @click="showBulkEditMemberModal"
+              v-b-tooltip.hover
+              ref="bulk-edit-members-button"
+              :title="$t('rooms.members.bulkEditUser',{numberOfSelectedUsers: selectedMembers.length})"
+            >
+              <i class="fas fa-users-cog"></i>
+            </b-button>
+            <!-- bulk remove member -->
+            <b-button
+              :disabled="isBusy"
+              variant="danger"
+              @click="showBulkRemoveMemberModal"
+              v-b-tooltip.hover
+              ref="bulk-remove-members-button"
+              :title="$t('rooms.members.bulkRemoveUser',{numberOfSelectedUsers: selectedMembers.length})"
+            >
+              <i class="fas fa-users-slash"></i>
+            </b-button>
+          </b-button-group>
       </div>
     </div>
 
@@ -155,6 +188,47 @@
       </div>
     </b-modal>
 
+    <!-- bulk edit users role modal -->
+    <b-modal
+      :static='modalStatic'
+      :busy="isLoadingAction"
+      ok-variant="success"
+      :cancel-title="$t('rooms.members.modals.edit.cancel')"
+      @ok="saveBulkEditMembers"
+      ref="bulk-edit-members-modal"
+      :no-close-on-esc="isLoadingAction"
+      :no-close-on-backdrop="isLoadingAction"
+      :hide-header-close="isLoadingAction"
+    >
+      <template v-slot:modal-title>
+        <span>
+        {{ $t('rooms.members.modals.edit.titleBulk', {numberOfSelectedUsers: selectedMembers.length}) }}
+          </span>
+      </template>
+      <template v-slot:modal-ok>
+        <b-spinner small v-if="isLoadingAction"></b-spinner>  {{ $t('rooms.members.modals.edit.save') }}
+      </template>
+      <div>
+        <b-form-group :state="fieldState('role')" :label="$t('rooms.members.modals.edit.role')">
+          <b-form-radio :state="fieldState('role')" v-model.number="bulkEditRole" name="some-radios" value="1">
+            <b-badge class="text-white" variant="success">{{ $t('rooms.members.roles.participant') }}</b-badge>
+          </b-form-radio>
+          <b-form-radio :state="fieldState('role')" v-model.number="bulkEditRole" name="some-radios" value="2">
+            <b-badge variant="danger">{{ $t('rooms.members.roles.moderator') }}</b-badge>
+          </b-form-radio>
+          <b-form-radio :state="fieldState('role')" v-model.number="bulkEditRole" name="some-radios" value="3">
+            <b-badge variant="dark">{{ $t('rooms.members.roles.co_owner') }}</b-badge>
+          </b-form-radio>
+
+          <template slot='invalid-feedback'><div v-html="fieldError('role')"></div></template>
+
+        </b-form-group>
+
+        <b-form-invalid-feedback :force-show="true"><div v-html="fieldError('users',true)"></div></b-form-invalid-feedback>
+
+      </div>
+    </b-modal>
+
     <!-- remove user modal -->
     <b-modal
       :busy="isLoadingAction"
@@ -169,14 +243,41 @@
       :hide-header-close="isLoadingAction"
     >
       <template v-slot:modal-title>
-        {{ $t('rooms.members.modals.delete.title') }}
+        {{ $t('rooms.members.modals.remove.title') }}
       </template>
       <template v-slot:modal-ok>
         <b-spinner small v-if="isLoadingAction"></b-spinner>  {{ $t('app.yes') }}
       </template>
-      <span v-if="deleteMember">
-        {{ $t('rooms.members.modals.delete.confirm',{firstname: deleteMember.firstname,lastname: deleteMember.lastname}) }}
+      <span v-if="removeMember">
+        {{ $t('rooms.members.modals.remove.confirm',{firstname: removeMember.firstname,lastname: removeMember.lastname}) }}
       </span>
+
+    </b-modal>
+
+    <!-- bulk remove user modal -->
+    <b-modal
+      :busy="isLoadingAction"
+      :static='modalStatic'
+      ok-variant="danger"
+      cancel-variant="dark"
+      :cancel-title="$t('app.no')"
+      @ok="confirmBulkRemoveMember"
+      ref="bulk-remove-members-modal"
+      :no-close-on-esc="isLoadingAction"
+      :no-close-on-backdrop="isLoadingAction"
+      :hide-header-close="isLoadingAction"
+    >
+      <template v-slot:modal-title>
+        {{ $t('rooms.members.modals.remove.titleBulk', {numberOfSelectedUsers: selectedMembers.length}) }}
+      </template>
+      <template v-slot:modal-ok>
+        <b-spinner small v-if="isLoadingAction"></b-spinner>  {{ $t('app.yes') }}
+      </template>
+      <span>
+        {{ $t('rooms.members.modals.remove.confirmBulk', {numberOfSelectedUsers: selectedMembers.length}) }}
+      </span>
+
+      <b-form-invalid-feedback :force-show="true"><div v-html="fieldError('users',true)"></div></b-form-invalid-feedback>
 
     </b-modal>
 
@@ -184,6 +285,7 @@
     <b-modal
       :busy="isLoadingAction"
       :static='modalStatic'
+      :ok-disabled="!newMember"
       ok-variant="success"
       :cancel-title="$t('rooms.members.modals.add.cancel')"
       @ok="saveNewMember"
@@ -276,18 +378,46 @@ export default {
       members: [], // list of all members
       createError: null, // error on adding new user as member
       editMember: null, // member to be edited
-      deleteMember: null, // member to be deleted
+      bulkEditRole: null, // role that will be applied to multiple members
+      removeMember: null, // member to be removed
       errors: {},
-      currentPage: 1
+      currentPage: 1,
+      selectedMembers: [],
+      displayedMembers: []
     };
   },
   methods: {
 
     /**
-     * Remove given member from member list
+     * select all checkbox toggled
+     * @param allSelected boolean select all rows
      */
-    removeMemberItem: function (member) {
-      this.members.splice(this.members.findIndex(item => item.id === member.id), 1);
+    onAllRowsSelected (allSelected) {
+      if (!allSelected) {
+        // If checkbox is now unchecked, remove all selections
+        this.$refs.membersTable.clearSelected();
+      } else {
+        // If checkbox is now checked, select all rows
+        this.$refs.membersTable.selectAllRows();
+      }
+    },
+
+    /**
+     * Row selection changed
+     * @param selectedRows selected rows
+     */
+    onRowsSelected (selectedRows) {
+      // set selected members to all selected table columns, exclude current user (prevent self edit/removed)
+      this.selectedMembers = selectedRows.filter(user => user.id !== this.currentUser.id);
+    },
+
+    /**
+     * Remove given members from member list, usable for both single removal or bulk removal
+     */
+    removeMemberItems: function (members) {
+      members.forEach(member => {
+        this.members.splice(this.members.findIndex(item => item.id === member), 1);
+      });
     },
 
     /**
@@ -316,8 +446,17 @@ export default {
      * @param member member object
      */
     showRemoveMemberModal: function (member) {
-      this.deleteMember = member;
+      this.removeMember = member;
       this.$refs['remove-member-modal'].show();
+    },
+
+    /**
+     * show modal to bulk remove members
+     * @param member member object
+     */
+    showBulkRemoveMemberModal: function () {
+      this.errors = {};
+      this.$refs['bulk-remove-members-modal'].show();
     },
 
     /**
@@ -329,14 +468,14 @@ export default {
 
       this.isLoadingAction = true;
 
-      Base.call('rooms/' + this.room.id + '/member/' + this.deleteMember.id, {
+      Base.call('rooms/' + this.room.id + '/member/' + this.removeMember.id, {
         method: 'delete'
       }).then(response => {
         // remove user entry from list
-        this.removeMemberItem(this.deleteMember);
+        this.removeMemberItems([this.removeMember.id]);
       }).catch((error) => {
         if (error.response.status === env.HTTP_GONE) {
-          this.removeMemberItem(this.deleteMember);
+          this.removeMemberItems([this.removeMember.id]);
         }
         Base.error(error, this.$root);
       }).finally(() => {
@@ -345,6 +484,41 @@ export default {
       });
     },
     /**
+     * Remove multiple members
+     * @param bvModalEvt
+     */
+    confirmBulkRemoveMember: function (bvModalEvt) {
+      // prevent modal from closing
+      bvModalEvt.preventDefault();
+
+      this.isLoadingAction = true;
+
+      const toBeRemovedMembers = this.selectedMembers.map(user => user.id);
+
+      Base.call('rooms/' + this.room.id + '/member', {
+        method: 'delete',
+        data: { users: toBeRemovedMembers }
+      }).then(response => {
+        // remove user entry from list
+        this.removeMemberItems(toBeRemovedMembers);
+        this.$refs['bulk-remove-members-modal'].hide();
+      }).catch((error) => {
+        // removing failed
+        if (error.response) {
+          // failed due to form validation errors
+          if (error.response.status === env.HTTP_UNPROCESSABLE_ENTITY) {
+            this.errors = error.response.data.errors;
+            return;
+          }
+        }
+        this.$refs['bulk-remove-members-modal'].hide();
+        Base.error(error, this.$root);
+      }).finally(() => {
+        this.isLoadingAction = false;
+      });
+    },
+
+    /**
      * show modal to edit user role
      * @param user user object
      */
@@ -352,6 +526,17 @@ export default {
       // Clone object to edit properties without displaying the changes in realtime in the members list
       this.editMember = _.cloneDeep(member);
       this.$refs['edit-member-modal'].show();
+    },
+
+    /**
+     * show modal to edit multiple users role
+     * @param user user object
+     */
+    showBulkEditMemberModal: function () {
+      // Clone object to edit properties without displaying the changes in realtime in the members list
+      this.errors = {};
+      this.bulkEditRole = null;
+      this.$refs['bulk-edit-members-modal'].show();
     },
 
     /**
@@ -365,23 +550,33 @@ export default {
     },
 
     /**
-     * save new user role
+     * Edit role of given members from member list for both single or bulk update
+     */
+    editMemberItems: function (members, role) {
+      members.forEach(member => {
+        this.members[this.members.findIndex(item => item.id === member)].role = role;
+      });
+    },
+
+    /**
+     *  save new user role
      */
     saveEditMember: function (bvModalEvt) {
       // prevent modal from closing
       bvModalEvt.preventDefault();
 
       this.isLoadingAction = true;
+      this.errors = {};
 
       Base.call('rooms/' + this.room.id + '/member/' + this.editMember.id, {
         method: 'put',
         data: { role: this.editMember.role }
       }).then(response => {
         // user role was saved
-        this.members[this.members.findIndex(item => item.id === this.editMember.id)].role = this.editMember.role;
+        this.editMemberItems([this.editMember.id], this.editMember.role);
       }).catch((error) => {
         if (error.response.status === env.HTTP_GONE) {
-          this.removeMemberItem(this.editMember);
+          this.removeMemberItems([this.editMember]);
         }
         Base.error(error, this.$root);
       }).finally(() => {
@@ -389,6 +584,43 @@ export default {
         this.isLoadingAction = false;
       });
     },
+
+    /**
+     * save new user roles for multiple members
+     * @param bvModalEvt
+     */
+    saveBulkEditMembers: function (bvModalEvt) {
+      // prevent modal from closing
+      bvModalEvt.preventDefault();
+
+      this.isLoadingAction = true;
+      this.errors = {};
+
+      const toBeEditedMembers = this.selectedMembers.map(user => user.id);
+
+      Base.call('rooms/' + this.room.id + '/member', {
+        method: 'put',
+        data: { role: this.bulkEditRole, users: toBeEditedMembers }
+      }).then(response => {
+        // user role was saved
+        this.editMemberItems(toBeEditedMembers, this.bulkEditRole);
+        this.$refs['bulk-edit-members-modal'].hide();
+      }).catch((error) => {
+        // editing failed
+        if (error.response) {
+          // failed due to form validation errors
+          if (error.response.status === env.HTTP_UNPROCESSABLE_ENTITY) {
+            this.errors = error.response.data.errors;
+            return;
+          }
+        }
+        Base.error(error, this.$root);
+        this.$refs['bulk-edit-members-modal'].hide();
+      }).finally(() => {
+        this.isLoadingAction = false;
+      });
+    },
+
     /**
      * add a user as a room member
      * @param bvModalEvt
@@ -455,6 +687,11 @@ export default {
       currentUser: state => state.session.currentUser
     }),
 
+    // amount of members that can be selected on the current page (user cannot select himself)
+    selectableMembers: function () {
+      return this.displayedMembers.filter(member => member.id !== this.currentUser.id).length;
+    },
+
     // check if new user input field is valid, local and server-side check
     newMemberValid: function () {
       if (this.newMember == null || this.newMember.id == null || this.fieldState('user') === false) { return false; }
@@ -510,6 +747,9 @@ export default {
           label: this.$t('rooms.members.actions'),
           thClass: 'action-column',
           tdClass: 'action-button'
+        });
+        fields.unshift({
+          key: 'selected'
         });
       }